Tidy up after merge
