+++ services/ui/src/app/components/tools/zones.tsx
@@ -143,6 +143,8 @@ export const Zones = ({ plotId, plotReady, forceUpdate }: ToolingProps) => {
           .attr("opacity", 0.5)
           .attr("style", "pointer-events: all")
           .style("cursor", "move")
+          .attr("stroke", "black")
+          .attr("stroke-width", 1)
           .datum(zone)
           .call(drag)
           .on("contextmenu", handleContextMenu);
